migrate pdp, landing-page, and plp to app router
