+++ packages/core/src/components/cms/RenderSections.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import {
   type ComponentType,
   type PropsWithChildren,
