+++ packages/core/src/components/templates/LandingPage/LandingPage.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import type { Locator } from '@vtex/client-cms'
 import { NextSeo, SiteLinksSearchBoxJsonLd } from 'next-seo'
 import type { ComponentType } from 'react'
@@ -101,46 +103,4 @@ export default function LandingPage({
       </PageProvider>
     </>
   )
-}
-
-export const getLandingPageBySlug = async (
-  slug: string,
-  previewData: Locator
-) => {
-  try {
-    if (storeConfig.cms.data) {
-      const cmsData = JSON.parse(storeConfig.cms.data)
-      const pageBySlug = cmsData['landingPage'].find((page: any) => {
-        slug === page.settings?.seo?.slug
-      })
-
-      if (pageBySlug) {
-        const landingPageData = await getPage<PageContentType>({
-          contentType: 'landingPage',
-          documentId: pageBySlug.documentId,
-          versionId: pageBySlug.versionId,
-        })
-
-        return landingPageData
-      }
-    }
-
-    const landingPageData = await getPage<PageContentType>({
-      ...(previewData?.contentType === 'landingPage'
-        ? previewData
-        : {
-            filters: {
-              filters: { 'settings.seo.slug': `/${slug}` },
-            },
-          }),
-      contentType: 'landingPage',
-    })
-    return landingPageData
-  } catch (error) {
-    if (error instanceof MissingContentError) {
-      return null
-    }
-
-    throw error
-  }
 }
