+++ packages/core/src/sdk/search/state.ts
@@ -1,5 +1,5 @@
 import { useCallback } from 'react'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/navigation'
 
 export const useApplySearchState = () => {
   const router = useRouter()
