+++ packages/core/src/sdk/ui/useOnClickOutside.ts
@@ -4,7 +4,7 @@ import { useEffect } from 'react'
 type Handler = (event: any) => void
 
 export default function useOnClickOutside<T extends HTMLElement = HTMLElement>(
-  ref: RefObject<T>,
+  ref: RefObject<T | null>,
   handler: Handler
 ) {
   useEffect(
